requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-chroma
langchain-community
langchain-huggingface
pypdf
python-dotenv
sentence-transformers
